pyproject.toml
[project]
name = "ai-agent-impact"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "anthropic>=0.49.0",
    "backoff>=2.2.1",
    "bs4>=0.0.2",
    "google-cloud-storage>=3.6.0",
    "google-cloud-aiplatform>=1.128.0",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.12",
    "langchain-google-genai>=3.2.0",
    "langchain-huggingface>=0.2.0",
    "langchain-ollama>=0.2.0",
    "langchain-openai>=0.3.21",
    "openai>=1.75.0",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rich>=14.2.0",
    "selenium>=4.38.0",
    "fastapi>=0.122.0",
    "datasets>=4.4.1",
    "uvicorn>=0.38.0",
    "langchain-aws>=0.2.31",
]

[dependency-groups]
analysis = [
    "matplotlib>=3.10.8",
    "patsy>=1.0.2",
    "seaborn>=0.13.2",
    "typer>=0.20.1",
    "scipy>=1.16.3",
    "statsmodels>=0.14.6",
]
dev = [
    "ipykernel>=7.1.0",
    "isort>=7.0.0",
    "pandas-stubs~=2.3.3",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "scipy-stubs~=1.16.3",
]

[tool.uv]
package = true

[tool.uv.workspace]
members = [
    "sandbox",
    "agent",
]

[project.scripts]
analysis = "experiments.analysis.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["experiments*", "common*", "agent*"]

